#97 [REFACTOR] 리뷰 작성에서 공통되는 컴포넌트 분리
